+++ tests/test_printer.py
@@ -42,7 +42,7 @@ class TestPrinter(unittest.TestCase):
         self.printer.action = Gtk.PrintOperationAction.EXPORT
         pdf = "Printer.test_print_records.pdf"
         self.printer.operation.set_export_filename(pdf)
-        records = [{"CALL": "MYCALL", "QSO_DATE": "24062017", "TIME_ON": "1519", "FREQ": "145.550", "MODE": "FM"}]
+        records = [{"id": 1, "CALL": "MYCALL", "QSO_DATE": "24062017", "TIME_ON": "1519", "FREQ": "145.550", "MODE": "FM", "RST_SENT": "59", "RST_RCVD": "57"}]
         result = self.printer.print_records(records)
         assert(result != Gtk.PrintOperationResult.ERROR)
         assert(result == Gtk.PrintOperationResult.APPLY)